requirements.txt
asgiref==3.2.5
Django==3.0.4
django-cleanup==4.0.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
